+++ projects/material-time-picker/src/lib/w-time-dialog/w-time-dialog.component.ts
@@ -1,11 +1,14 @@
 import { CLOCK_TYPE, ITime } from '../w-clock/w-clock.component';
 import { Component, Inject } from '@angular/core';
-import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MAT_DIALOG_DATA, MatDialogContent } from '@angular/material/dialog';
 import { MatDialogRef } from '@angular/material/dialog';
+import { WTimeComponent } from '../w-time/w-time.component';
 
 @Component({
-  styleUrls: ['./w-time-dialog.component.scss'],
-  templateUrl: './w-time-dialog.component.html'
+    styleUrls: ['./w-time-dialog.component.scss'],
+    templateUrl: './w-time-dialog.component.html',
+    standalone: true,
+    imports: [MatDialogContent, WTimeComponent]
 })
 export class WTimeDialogComponent {
   userTime: ITime;