+++ projects/material-time-picker/src/public_api.ts
@@ -7,4 +7,4 @@ export { WTimeComponent } from './lib/w-time/w-time.component';
 export { WTimeDialogComponent } from './lib/w-time-dialog/w-time-dialog.component';
 export { MaterialTimePickerComponent } from './lib/material-timepicker/material-timepicker.component';
 
-export * from './lib/material-time-picker.module';
+